YouTube_Data_Harvesting_and_Warehousing/requirements.txt
google_api_python_client==2.87.0
mysql_connector_repackaged==0.3.1
numpy==1.24.2
pandas==1.5.3
pymongo==4.3.3
requests==2.28.2
streamlit>=1.20.0
